Revert whitespace/include changes that aren't needed; move plSrtFileReader to plAudioCore to fix unity build issues; apply code suggestions from Adam
